+++ src/Controller/ClapController.php
@@ -2,18 +2,8 @@
 
 namespace App\Controller;
 
-use App\AutoMapping;
-use App\Request\CreateClapRequest;
-use App\Request\DeleteRequest;
-use App\Request\GetClientRequest;
-use App\Request\GetEntityRequest;
-use App\Request\UpdateClapRequest;
 use App\Service\ClapService;
 use App\Validator\ClapValidateInterface;
-use AutoMapperPlus\AutoMapper;
-use AutoMapperPlus\Configuration\AutoMapperConfig;
-use AutoMapperPlus\Exception\UnregisteredMappingException;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -21,24 +11,20 @@ use Symfony\Component\Routing\Annotation\Route;
 class ClapController extends BaseController
 {
     private $clapService;
-    private $autoMapping;
 
     /**
      * ClapController constructor.
      * @param $clapService
      */
-    public function __construct(ClapService $clapService,AutoMapping $autoMapping)
+    public function __construct(ClapService $clapService)
     {
         $this->clapService = $clapService;
-        $this->autoMapping=$autoMapping;
     }
 
     /**
      * @Route("/claps", name="createClap",methods={"POST"})
      * @param Request $request
-     * @param ClapValidateInterface $clapValidate
-     * @return JsonResponse|Response
-     * @throws UnregisteredMappingException
+     * @return
      */
     public function create(Request $request, ClapValidateInterface $clapValidate)
     {
@@ -50,18 +36,16 @@ class ClapController extends BaseController
             $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
             return $resultResponse;
         }
-        $data = json_decode($request->getContent(), true);
-        $request=$this->autoMapping->map(\stdClass::class,CreateClapRequest::class,(object)$data);
+        //
+
         $result = $this->clapService->create($request);
-        return $this->response($result, self::CREATE);
+        return $this->response($result, self::CREATE, "Clap");
     }
 
     /**
      * @Route("/clap/{id}", name="updateClap",methods={"PUT"})
      * @param Request $request
-     * @param ClapValidateInterface $clapValidate
-     * @return JsonResponse|Response
-     * @throws UnregisteredMappingException
+     * @return
      */
     public function update(Request $request, ClapValidateInterface $clapValidate)
     {
@@ -72,57 +56,59 @@ class ClapController extends BaseController
             $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
             return $resultResponse;
         }
-        $id=$request->get('id');
-        $data = json_decode($request->getContent(), true);
-        $request=$this->autoMapping->map(\stdClass::class,UpdateClapRequest::class,(object)$data);
-        $request->setId($id);
         $result = $this->clapService->update($request);
-        return $this->response($result, self::UPDATE);
+        return $this->response($result, self::UPDATE, "Clap");
     }
 
     /**
-     * @Route("/clap/{id}", name="deleteClap",methods={"DELETE"})
+     *  @Route("/clap/{id}", name="deleteClap",methods={"DELETE"})
      * @param Request $request
-     * @return JsonResponse
+     * @return
      */
-    public function delete(Request $request)
+    public function delete(Request $request, ClapValidateInterface $clapValidate)
     {
-        $request=new DeleteRequest($request->get('id'));
+        $validateResult = $clapValidate->clapValidator($request, 'delete');
+        if (!empty($validateResult))
+        {
+            $resultResponse = new Response($validateResult, Response::HTTP_OK, ['content-type' => 'application/json']);
+            $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
+            return $resultResponse;
+        }
         $result = $this->clapService->delete($request);
-        return $this->response($result, self::DELETE);
+        return $this->response($result, self::DELETE,"Clap");
+
     }
+
 
     /**
      * @Route("/clapsentity/{entity}/{row}",name="getEntityClap",methods={"GET"})
      * @param Request $request
-     * @return JsonResponse
+     * @return
      */
-    public function getEntityClap(Request $request)
+    public function getEntityclap(Request $request)
     {
-        $request=new GetEntityRequest($request->get('entity'),$request->get('row'));
         $result = $this->clapService->getEntityClap($request);
-        return $this->response($result,self::FETCH);
+        return $this->response($result,self::FETCH,"Clap");
     }
 
     /**
      * @Route("/clapsclient/{client}", name="getClientClaps",methods={"GET"})
      * @param Request $request
-     * @return JsonResponse
+     * @return
      */
     public function getClientClap(Request $request)
     {
-        $request=new GetClientRequest($request->get('client'));
         $result = $this->clapService->getClientClap($request);
-        return $this->response($result,self::FETCH);
+        return $this->response($result,self::FETCH,"Clap");
     }
-
     /**
      * @Route("/claps",name="getAllClap",methods={"GET"})
-     * @return JsonResponse
+     * @param Request $request
+     * @return
      */
-    public function getAll()
+    public function getAll(Request $request)
     {
-        $result = $this->clapService->getAll();
-        return $this->response($result,self::FETCH);
+        $result = $this->clapService->getAll($request);
+        return $this->response($result,self::FETCH,"Clap");
     }
 }
