+++ src/Controller/FavoriteController.php
@@ -2,11 +2,6 @@
 
 namespace App\Controller;
 
-use App\AutoMapping;
-use App\Request\ByIdRequest;
-use App\Request\CreateFavoriteRequest;
-use App\Request\DeleteRequest;
-use App\Request\UpdateFavoriteRequest;
 use App\Service\FavoriteService;
 use App\Validator\FavoriteValidateInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -17,22 +12,19 @@ use Symfony\Component\Routing\Annotation\Route;
 class FavoriteController extends BaseController
 {
     private $favoriteService;
-    private $autoMapping;
 
-    public function __construct(FavoriteService $favoriteService,AutoMapping $autoMapping)
+    public function __construct(FavoriteService $favoriteService)
     {
-        $this->autoMapping=$autoMapping;
         $this->favoriteService=$favoriteService;
     }
-
     /**
      * @Route("/favorites", name="createFavorite",methods={"POST"})
      * @param Request $request
-     * @param FavoriteValidateInterface $favoriteValidate
-     * @return JsonResponse|Response
+     * @return
      */
     public function create(Request $request, FavoriteValidateInterface $favoriteValidate)
     {
+        //Validation
         $validateResult = $favoriteValidate->favoriteValidator($request, 'create');
         if (!empty($validateResult))
         {
@@ -40,18 +32,15 @@ class FavoriteController extends BaseController
             $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
             return $resultResponse;
         }
-        $data = json_decode($request->getContent(), true);
-        $data=(object)$data;
-        $request=$this->autoMapping->map(\stdClass::class,CreateFavoriteRequest::class,$data);
+        //
+
         $result = $this->favoriteService->create($request);
-        return $this->response($result, self::CREATE);
+        return $this->response($result, self::CREATE, "Favorite");
     }
 
-    /*
-     * @Route("/favorite/{id}" , name="updateFavorite", methods={"PUT"})
+    /**@Route("/favorite/{id}", name="updateFavorite",methods={"PUT"})
      * @param Request $request
-     * @param FavoriteValidateInterface $favoriteValidate
-     * @return JsonResponse|Response
+     * @return
      */
     public function update(Request $request, FavoriteValidateInterface $favoriteValidate)
     {
@@ -62,49 +51,49 @@ class FavoriteController extends BaseController
             $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
             return $resultResponse;
         }
-        $id=$request->get('id');
-
-        $data = json_decode($request->getContent(), true);
-        $request=$this->autoMapping->map(\stdClass::class,UpdateFavoriteRequest::class,(object)$data);
-        $request->setId($id);
         $result = $this->favoriteService->update($request);
-        return $this->response($result, self::UPDATE);
+        return $this->response($result, self::UPDATE, "Favorite");
     }
 
     /**
-     * @Route("/favorite/{id}", name="deleteFavorite",methods={"DELETE"})
+     *  @Route("/favorite/{id}", name="deleteFavorite",methods={"DELETE"})
      * @param Request $request
-     * @param FavoriteValidateInterface $favoriteValidate
-     * @return JsonResponse|Response
+     * @return
      */
     public function delete(Request $request, FavoriteValidateInterface $favoriteValidate)
     {
-        $request=new DeleteRequest($request->get('id'));
+        $validateResult = $favoriteValidate->favoriteValidator($request, 'delete');
+        if (!empty($validateResult))
+        {
+            $resultResponse = new Response($validateResult, Response::HTTP_OK, ['content-type' => 'application/json']);
+            $resultResponse->headers->set('Access-Control-Allow-Origin', '*');
+            return $resultResponse;
+        }
         $result = $this->favoriteService->delete($request);
-        return $this->response($result, self::DELETE);
+        return $this->response($result, self::DELETE,"Favorite");
 
     }
 
+
     /**
-     * @Route("/favoritesclient/{id}",name="getClientFavorite",methods={"GET"})
+     * @Route("/favoritesClient",name="getClientFavorite",methods={"GET"})
      * @param Request $request
-     * @return JsonResponse
+     * @return
      */
     public function getClientFavorite(Request $request)
     {
-        $request=new ByIdRequest($request->get('id'));
         $result = $this->favoriteService->getClientFavorite($request);
-        return $this->response($result,self::FETCH);
+        return $this->response($result,self::FETCH,"Favorite");
     }
 
     /**
      * @Route("/favorites",name="getAllFavorite",methods={"GET"})
-     * @return JsonResponse
+     * @return
      */
     public function getAll()
     {
         $result = $this->favoriteService->getAll();
-        return $this->response($result,self::FETCH);
+        return $this->response($result,self::FETCH,"Favorite");
     }
 
 }
